refactor: split advancedActions into individual card files
Move advanced action card definitions from bundled color files
(red.ts, blue.ts, etc.) into individual files per card to minimize
merge conflicts when multiple developers work on cards in parallel.

New structure:
- advancedActions/{red,blue,white,green}/
  - Each color has its own directory with individual card files
  - index.ts re-exports cards and provides the color's record
- advancedActions/dual/
  - Dual-color cards that can be powered by two colors
- Bolt cards moved to their respective color directories
  (fire-bolt → red, ice-bolt → blue, swift-bolt → white,
   crushing-bolt → green)

Adding a new card now only requires:
1. Create a new file (e.g., newCard.ts)
2. Add exports and import to the color's index.ts
